+++ ChangeLog.md
@@ -82,6 +82,11 @@ V0.0.13 changelog
 
   Refactor the time code output function to reduce code duplication
 
+添加 Add:
+- 添加`get_audio_sample_count` 方法用于正确输出TC时间戳下的音频采样数, 解决issue [#9](https://github.com/OwenYou/dftt_timecode/issues/9)
+
+  Add `get_audio_sample_count` method for correctly outputting the count of audio samples at TC timestamps,solve issue [#9](https://github.com/OwenYou/dftt_timecode/issues/9)
+
 弃用 Deprecate:
 - 使用`functools.singledispatchmethod` 代替 `dispatch.InstanceMethodDispatch`